+++ spider-link/src/main/java/com/example/spiderlink/config/SpiderLinkAutoConfiguration.java
@@ -1,10 +1,7 @@
 package com.example.spiderlink.config;
 
-import com.example.spiderlink.config.InternalApiInterceptor;
-import com.example.spiderlink.config.WebMvcConfig;
 import com.example.spiderlink.domain.loglevel.LogLevelApplier;
 import com.example.spiderlink.domain.management.ManagementReloadCommandHandler;
-import com.example.spiderlink.domain.management.ManagementReloadHttpController;
 import com.example.spiderlink.domain.management.executor.LogAdditivityExecutor;
 import com.example.spiderlink.domain.management.executor.LogLevelExecutor;
 import com.example.spiderlink.domain.management.executor.ManagementExecutor;
@@ -15,8 +12,6 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication.Type;
 import org.springframework.boot.autoconfigure.jdbc.JdbcTemplateAutoConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -34,10 +29,13 @@ import org.springframework.jdbc.core.JdbcTemplate;
  *   <li>{@link LogLevelExecutor} — {@code gubun=log_config_level} 관리 명령 처리기</li>
  *   <li>{@link LogAdditivityExecutor} — {@code gubun=log_config_additivity} 관리 명령 처리기</li>
  *   <li>{@link ManagementReloadCommandHandler} — TCP {@code MANAGEMENT_RELOAD} 명령 핸들러</li>
- *   <li>{@link ManagementReloadHttpController} — HTTP {@code POST /api/management/reload} 핸들러</li>
- *   <li>{@link InternalApiInterceptor} — {@code /api/internal/**}, {@code /api/management/**} IP 접근 제어 (Servlet WAS 전용)</li>
- *   <li>{@link WebMvcConfig} — {@link InternalApiInterceptor} MVC 인터셉터 등록 (Servlet WAS 전용)</li>
  * </ul>
+ *
+ * <p>Servlet WAS 전용 빈({@code ManagementReloadHttpController}, {@code InternalApiInterceptor},
+ * {@code WebMvcConfig})은 {@link SpiderLinkServletAutoConfiguration}에서 별도 관리된다.
+ * spring-webmvc가 없는 TCP 전용 WAS에서 이 클래스 introspect 시 발생하는
+ * {@code NoClassDefFoundError}를 방지하기 위해 분리되었다.</p>
+ *
  */
 // JdbcTemplateAutoConfiguration 이후에 실행해야 JdbcTemplate 빈이 이미 등록된 상태에서
 // @ConditionalOnBean(JdbcTemplate.class) 조건을 올바르게 평가할 수 있다
@@ -112,50 +110,4 @@ public class SpiderLinkAutoConfiguration {
         return new ManagementReloadCommandHandler(executors);
     }
 
-    /**
-     * HTTP 관리 명령 수신 컨트롤러 빈.
-     *
-     * <p>Admin → HTTP WAS 경로의 Reload 요청을 처리한다.
-     * TCP의 {@link ManagementReloadCommandHandler}와 동일한 {@link ManagementExecutor} 패턴을 사용한다.
-     * web-application-type=none 인 TCP 전용 WAS에서는 빈이 생성되더라도 Spring MVC가 없으므로
-     * HTTP 엔드포인트로 노출되지 않아 부작용이 없다.</p>
-     *
-     * @param executors Spring 컨텍스트에 등록된 모든 ManagementExecutor 구현체 목록
-     */
-    @Bean
-    @ConditionalOnBean(ManagementExecutor.class)
-    @ConditionalOnMissingBean
-    public ManagementReloadHttpController managementReloadHttpController(
-            List<ManagementExecutor> executors) {
-        return new ManagementReloadHttpController(executors);
-    }
-
-    /**
-     * 내부 API IP 접근 제어 인터셉터 빈 (Servlet WAS 전용).
-     *
-     * <p>standalone 실행 시에는 컴포넌트 스캔으로 이미 등록되므로 {@code @ConditionalOnMissingBean}으로
-     * 중복 등록을 방지한다. 라이브러리 JAR로 포함되는 경우에만 이 빈이 실제로 생성된다.</p>
-     *
-     * @param allowedIps {@code internal-api.allowed-ips} 프로퍼티 (기본값: loopback 3종)
-     */
-    @Bean
-    @ConditionalOnWebApplication(type = Type.SERVLET)
-    @ConditionalOnMissingBean
-    public InternalApiInterceptor internalApiInterceptor(
-            @Value("${internal-api.allowed-ips:127.0.0.1,0:0:0:0:0:0:0:1,::1}") List<String> allowedIps) {
-        return new InternalApiInterceptor(allowedIps);
-    }
-
-    /**
-     * MVC 인터셉터 등록 설정 빈 (Servlet WAS 전용).
-     *
-     * <p>{@link InternalApiInterceptor}를 {@code /api/internal/**}, {@code /api/management/**} 경로에
-     * 등록한다. standalone 실행 시에는 컴포넌트 스캔으로 이미 등록되므로 중복 등록을 방지한다.</p>
-     */
-    @Bean
-    @ConditionalOnWebApplication(type = Type.SERVLET)
-    @ConditionalOnMissingBean(WebMvcConfig.class)
-    public WebMvcConfig spiderLinkWebMvcConfig(InternalApiInterceptor internalApiInterceptor) {
-        return new WebMvcConfig(internalApiInterceptor);
-    }
 }
