+++ spider-link/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,2 @@
 com.example.spiderlink.config.SpiderLinkAutoConfiguration
+com.example.spiderlink.config.SpiderLinkServletAutoConfiguration
